exercises 4.17-21 done

+++ Part4/bloglist/controllers/blogs.js
@@ -1,53 +1,72 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
-blogsRouter.get('/', (request, response) => {
-    Blog
-        .find({})
-        .then(blogs => {
-            response.json(blogs)
-        })
+blogsRouter.get('/', async (request, response) => {
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1, id: 1 })
+    response.json(blogs)
 })
 
-blogsRouter.post('/', (request, response) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', async (request, response, next) => {
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+        title: body.title,
+        author: body.author,
+        url: body.url,
+        likes: body.likes,
+        user: user._id
+    })
+
     if (!blog.title && !blog.url) {
         return response.status(400).send('Bad Request')
     }
 
     if (!blog.likes) blog.likes = 0
-    blog
-        .save()
-        .then(result => {
-            response.status(201).json(result)
-        })
+    
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog)
 })
-
-// above get and post uses promises, below put and delete uses async/await
 
 blogsRouter.put('/:id', async (request, response, next) => {
     const body = request.body
-
     const blog = {
         title: body.title,
         author: body.author,
         url: body.url,
         likes: body.likes
     }
-
     try {
         const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
         response.json(updatedBlog.toJSON())
     } catch (exception) {
         next(exception)
     }
-
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
+        const blog = await Blog.findById(request.params.id)
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        if (blog.user.toString() === user._id.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            response.status(401).send({ error: 'user id does not match' })
+        }
     } catch (exception) {
         next(exception)
     }
